Implement Clone for LinearLayout and Chains
Added Clone implementation for LinearLayout and the ViewGroup Chain.
This will allow for the use of LinearLayouts that are backed by the
Chain ViewGroup to be used with the ViewEnum macro. Also modified the
ViewEnum macro to explicitly call Clone(). Previously we used structures
that implemented Copy. Since cloning Chains is non-trivial we need to
call clone explicitly.